Use uintxx_t for capacity and pageSize
This makes it more explicit what the range is. In addition, on smaller
architectures like AVR, size_t is 16-bit which is insufficient to store
the capacity of 512kbit and larger EEPROMs.

+++ src/EEPROM_CAT25.cpp
@@ -243,11 +243,11 @@ size_t EEPROM_CAT25::writeOrUpdatePage(bool update, const uint32_t address, cons
   return(length);
 }
 
-size_t EEPROM_CAT25::capacity() {
+uint32_t EEPROM_CAT25::capacity() {
   return this->_capacity;
 }
 
-size_t EEPROM_CAT25::pageSize() {
+uint16_t EEPROM_CAT25::pageSize() {
   return this->_pageSize;
 }
